Add default org lookup against the backend if not locally configured by user (#19324)
Essentially a copy of https://github.com/pulumi/pulumi/pull/19249 but
with commits re-organized for readability.
This work is part of Project Hockeystick
(https://github.com/orgs/pulumi/projects/274)

This PR updates the call sites for workspace.GetBackendConfigDefaultOrg
to a wrapper `backend.GetDefaultOrg` that includes a fallback lookup to
the backend for default org. If the backend does not support the
GetDefaultOrg API, or if it has no opinion, `backend.GetDefaultOrg` will
preserve existing behavior by returning an empty string for call sites
to evaluate appropriately.

It additionally introduces a new call site: cloudBackend.ListStacks` --
This upfront call will reduce the number of API calls made in
`cloudBackendReference.String()` to check the workspace against the
default organization when eliding stack route, reducing performance
impact for large scale organizations.