Allow for hyphens in function names (#19352)
This pull request introduces a new schema definition for function types.

Currently, function tokens may not have any hyphens in their final part.
For most existing functions, this is fine, since we habitually do not
create functions that have hyphens in them.

However, for provider-level Methods, which are of the token format
`pulumi:providers/packageName/functionName`, sometimes the provider has
a hyphen. Famous example: `terraform-provider` - the base name for our
provider parameterization.

Rather than widening the regex scope for all tokens, this pull request
adds a new `functionToken` to the metaschema with an additional allowed
`-` in the regex pattern.

Adds tests to confirm this is permissible for all function tokens, as
well as the special provider methods.

Fixes https://github.com/pulumi/pulumi/issues/19274.