Remodel `DependencyGraph`s to use pointer equality
The `DependencyGraph` type in `pkg/resource/graph` provides several
operations for computing dependencies between resources in a snapshot.
Presently, many of these operations use URNs and IDs as the keys for
finding nodes in the graph. This makes sense as in any given Pulumi
snapshot, dependencies are encoded using URNs (parent/child,
dependencies, property dependencies and deleted-with) and IDs (provider
references). Unfortunately, neither URNs nor IDs are guaranteed to be
truly unique in a snapshot, and so several operations have to use other
heuristics to work correctly.

While URNs and IDs aren't perfectly unique, each resource object _will_
have a unique address in memory. Relying on this however, requires that
all callers of `DependencyGraph` methods are passing objects that are
exactly those in the graph itself. Checking the current caller set, this
does appear to be the case. As a result, we can rework the
implementation of `DependencyGraph` to use pointer equality checks in
many places, producing a more accurate and heuristic-free picture of any
given dependency graph.

An added benefit of using pointer equality is that we can safely combine
dependency graphs that represent different views of the same set of
resources. For example, a set of resources from a _program_ with a set
from a _state snapshot_. This is exactly what we need to better
understand dependencies when `--run-program` is in play with e.g.
`destroy` -- in such cases, computing what must be deleted requires
traversing a graph that contains the union of registered resources and
base snapshot resources. A dependency graph that can (safely) contain
all resources and correctly identify relationships between them is a
useful tool to have here.

While we are here, we also fix some performance issues that have crept
in as the implementation has evolved so that both graph construction and
amortized operations on the graph perform better and (for the most part)
avoid undesirable quadratic-time behaviour. The recently added
`TestPerf` test illustrates the improvements -- here we are operating
over a graph of `n` resources where each resource `R_i` is a child of
`R_{i-1}`. For each operation, we perform it on the first resource in
the graph (`R_0` -- `low`), the middle resource in the graph (`mid` --
`R_{n/2}`), and the last resource in the graph (`high` -- `R_n`). For
each operation we expect certain characteristics -- e.g. finding all
the children of the `low` resource should be more expensive than finding
those of `high`, since `high` is the last resource in the graph and thus
can't have any children, and so on.

* Old code -- 10,000 resources; 5,000 iterations per operation

```
NewDependencyGraph (3.87s)
DependenciesOf
  low (0.00s)
  mid (0.14s)
  high (0.28s)
ParentsOf
  low (0.05s)
  mid (0.45s)
  high (0.41s)
ChildrenOf
  low (0.25s)
  mid (0.13s)
  high (0.00s)
TransitiveDependenciesOf
  low (5.19s)
  mid (9.18s)
  high (13.02s)
Contains
  low (0.00s)
  mid (0.00s)
  high (0.00s)
DependingOn
  low (2.90s)
  mid (1.43s)
  high (0.00s)
DependingOnWithChildren
  low (7.57s)
  mid (3.72s)
  high (0.00s)
OnlyDependsOn
  low (10.32s)
  mid (8.42s)
  high (5.89s)
Total (73.24s)
```

* Old code -- 25,000 resources; 5,000 iterations per operation

```
NewDependencyGraph (26.84s)
DependenciesOf
  low (0.00s)
  mid (0.24s)
  high (0.59s)
ParentsOf
  low (0.15s)
  mid (0.86s)
  high (1.41s)
ChildrenOf
  low (0.59s)
  mid (0.28s)
  high (0.00s)
TransitiveDependenciesOf
  low (11.89s)
  mid (22.01s)
  high (32.62s)
Contains
  low (0.00s)
  mid (0.00s)
  high (0.00s)
DependingOn
  low (7.48s)
  mid (3.68s)
  high (0.00s)
DependingOnWithChildren
  low (17.83s)
  mid (8.89s)
  high (0.00s)
OnlyDependsOn
  low (23.31s)
  mid (19.98s)
  high (13.36s)
Total (192.03s)
```

* New code -- 10,000 resources; 5,000 iterations per operation

```
NewDependencyGraph (0.02s)
DependenciesOf
  low (0.00s)
  mid (0.00s)
  high (0.00s)
ParentsOf
  low (0.00s)
  mid (0.31s)
  high (0.61s)
ChildrenOf
  low (1.21s)
  mid (0.54s)
  high (0.00s)
TransitiveDependenciesOf
  low (0.00s)
  mid (6.72s)
  high (13.90s)
Contains
  low (0.00s)
  mid (0.00s)
  high (0.00s)
DependingOn
  low (0.88s)
  mid (0.46s)
  high (0.00s)
DependingOnWithChildren
  low (9.16s)
  mid (4.62s)
  high (0.00s)
OnlyDependsOn
  low (9.20s)
  mid (5.53s)
  high (1.01s)
Total (54.20s)
```

* New code -- 25,000 resources; 5,000 iterations per operation

```
NewDependencyGraph (0.04s)
DependenciesOf
  low (0.00s)
  mid (0.00s)
  high (0.00s)
ParentsOf
  low (0.00s)
  mid (0.79s)
  high (1.82s)
ChildrenOf
  low (3.34s)
  mid (1.42s)
  high (0.00s)
TransitiveDependenciesOf
  low (0.00s)
  mid (13.70s)
  high (28.89s)
Contains
  low (0.00s)
  mid (0.00s)
  high (0.00s)
DependingOn
  low (2.60s)
  mid (1.42s)
  high (0.00s)
DependingOnWithChildren
  low (22.23s)
  mid (10.89s)
  high (0.00s)
OnlyDependsOn
  low (22.06s)
  mid (12.98s)
  high (2.77s)
Total (124.97s)
```

We can see that construction of a `NewDependencyGraph` has stopped being
quadratic/poorly performing and is now very fast, while the amortized
expensive operations such as `TransitiveDependenciesOf` and
`OnlyDependsOn` have slightly better constant factors also. Note that,
for the most part, we don't expect to see stacks with tens of thousands
of resources (these perform badly for other reasons), but it's good to
know we could work acceptably in these scenarios.