Respect customized Node package names in Node program generator
Packages may override Node imports to be @pulumiverse/scaleway instead of @pulumi/scaleway. Before the change the code
respected this for resources but not for invokes, and if invokes were used generated invalid Node programs. This is
fixed by reusing the code that handled resources to cover invokes as well.

Fixes #19409