Update tests/integration/go/parameterized/go.mod deps (#19413)
This change updates `tests/integration/go/parameterized/go.mod` to use
the latest versions of `golang.org/x/crypto` and `golang.org/x/net` to
address the last few dependabot alerts.

Additionally, this change makes it possible for dependabot to update
these dependencies automatically in the future. It previously couldn't
be updated automatically because the `go.mod` refers to a generated SDK
that doesn't exist. Now, we add those references programmatically as
part of running the test.