Support Resource References (#19226)
Similar to the Node.js implementation in
https://github.com/pulumi/pulumi/pull/18885

When we encounter an input or output that is a subclass of `Resource`,
we look up its type token. This is available under the `pulumi_type`
properties for SDKs built with
https://github.com/pulumi/pulumi/pull/19221. To retrieve the version,
download URL and parameterization for a Pulumi package, we load the
`pulumi-plugin.json` file. These packages are tracked as dependencies
and returned in the schema so that codegen can generate the correct
language specific dependencies.