Ensure we don’t introduce `toolchain` directives in the go.mod files
CI runs a matrix of Go versions, and we want to ensure we’re always building and testing with that version. If we have `toolchain` directives in the `go.mod` files, we’d use those versions instead, defeating the purpose of the matrix.

See https://github.com/pulumi/pulumi/pull/19118#discussion_r2071491860