cache intermediate results for type conversions
Type conversions are mutually recursive and can exhibit bad runtime
behaviour (probably O(2^n) if I'm guessing right) in some cases.  With
large schemas this means we essentially never finish the conversion.

To fix this we can use dynamic programming techniques, and cache the
intermediate results, which helps cut down the runtime for this
significantly.  Locally this brings the resourcedocgen command from
https://github.com/pulumi/pulumi-java/issues/1772, from never
finishing (or at least not before my patience runs out after running
it overnight) to completing in ~3 minutes, after including this change
in the Java language runtime.