Add failure case to test pkg parser (#19425)
I called a file something like `example-failure.json` and it caused this
particular function to get stuck in a loop. I didn't notice for a few
days because of (A) flakes and (B) the output not mentioning what had
gone wrong. This PR raises a specific error if an entire filename is
consumed without successfully parsing a semver version.