Adds FeatureFlagged Copilot assisted "Explain" option when Previewing Updates (#19138)
Fixes https://github.com/pulumi/pulumi.ai/issues/1789

Backend implemented in https://github.com/pulumi/pulumi.ai/pull/1787,
now merged.

> [!WARNING]
> This feature is not ready for general use yet and is only active when
`--copilot`, a hidden flag, is provided.

Looks like this:

```
Do you want to proceed?
> yes
  no
  details
  explain ✨
```

Explain will invoke copilot to give a summary of the update.

### Main code changes:

1. **Add a new API handler** to call Copilot's "explainPreview" skill.
2. **Tweak the logic around exposing whether Copilot is enabled for an
org**. Whether it is enabled is established by calling `apply`, the
explain flow is a parallel path, (unlike "summarizeUpdate") so we need
to set some state that explain can check (`copilotEnabledForProject`).
3. **Add explain option to the list of options to select**. Here we
intro a new `Explainer` interface to keep the pkg/backend layer a bit
generic. (DIY backends do not support this feature for now).

### Visibility of the `explain` option

It will appear if

1. You provide the temporary, hidden, `--copilot` flag to turn this
feature on (will be removed in a later release)
1. The stack is hosted on app.pulumi.com
2. The org the stack belongs to has Copilot enabled for members or
admins (and the user is an admin in the latter case).
3. You run either of `pulumi up`, `destroy` or `refresh`.

Screenshots


![image](https://github.com/user-attachments/assets/fd9dcd55-6c73-4377-8da5-055f0b10dcdb)

---------

Co-authored-by: Mikhail Shilkov <github@mikhail.io>