make namespaces we get from git repos always lowercase (#19441)
Namespaces are only allowed to be lowercase. For git repos we take them
from the URL, which is allowed to have uppercase characters and is
usually not case sensitive. Automatically lowercase the namespace, so
this always works and doesn't confuse users that use components from Git
repos that don't have an explicit namespace set.

E.g. someone might use `pulumi package add
github.com/Pulumi/xyz-component`, which should just use the lowercase
`pulumi` (as that's what the URL is pointing to because of case
insensitivity), instead of `Pulumi`.