Add some NodeJS request types (#19437)
I _think_ this should be an absolute no-op, as these types don't seem to
escape these functions.

We actually have types for the things we expect these to be, and if
we're wrong, everything will still fail in exactly the same way as
before this PR. This change should just mean that the LSP is a bit more
useful.