Fix the return type of the PCL `call` intrinsic
Method calls (represented by the `Call` gRPC call on the provider
interface and the `call` intrinsic in PCL) always return outputs -- we
don't support "plain" (e.g. `Promise` in NodeJS, `Task` in Python) calls
as we do for invokes. When the `call` PCL intrinsic was introduced, we
incorrectly gave it a plain type, which happened to work for the NodeJS
and Python conformance tests we introduced due to those languages'
support for automatic lifting of output properties. For instance, in
NodeJS, the below is fine, even though `component1.identity()` returns
an `Output`:

```typescript
export const from_identity = component1.identity().result;
```

Under the hood, "lifting" means that this is desugared to an `apply`:

```typescript
export const from_identity = component1.identity().apply(call => call.result);
```

This doesn't fly in languages like Java and Go though, where we don't
have lifting. This change fixes the type of the `call` intrinsic to
correctly report that it returns an output. Doing so means that program
code generation can proceed correctly in all languages.

Part of pulumi-java#262