Respect customized Node package names in Node program generator (#19410)
Packages may override Node imports to be @pulumiverse/scaleway instead
of @pulumi/scaleway. Before the change the code respected this for
resources but not for invokes, and if invokes were used generated
invalid Node programs. This is fixed by reusing the code that handled
resources to cover invokes as well.

Fixes https://github.com/pulumi/pulumi/issues/19409