Add MainContext to allow for provider cancellation (#19445)
`provider.Main` currently forces `context.Background()`, but this is
better left up to the caller. For example if I want to run a test
provider in-process for 1 second, I should be able to do that by passing
a context with a timeout.

For backward compatibility we keep `Main` as-is and introduce
`MainContext` alongside it.