Add `--run-program` to `preview`/`up` (#19449)
This adds `--run-program` to `pulumi preview` and `pulumi up`. This only
has an effect when `--refresh` is also used, but when set will use the
new refresh system as part of step generation, rather than refreshing
everything up-front based on state.

This has a _small_ difference that it _doesn't_ refresh resources that
are being deleted. That should probably be deemed a bug, but might be
we're ok shipping this as is first.

---------

Co-authored-by: Thomas Gummerer <t.gummerer@gmail.com>