Sdk changes required for scalar returns from providers
First in a series of changes to add scalar returns to method calls in providers.

The code for a scalar value in the schema (pcl spec) already exists for
the ReturnType field, this change adds the nodejs sdk and codegen
required to leverage that.

When it is all done it will work like so:

    1. Sdk was modified to have symmetric calls to the invoke* family of functions which handle returning single/scalar or objects as either output or plain values (promise).
    1. Codegen detects the type of ReturnType field is not ObjectType, and
    generates the appropriate sdk api code for the function call to
    return a scalar.
    1. Single keyed PropertyMaps are returned from the provider for the
       scalar.
    1. The generated code appropriately unmarshals the single value and
       returns from the call in the program's sdk code.

This PR is sdk only and will require a release before the programgen
code ships.