Support enum references
When we encounter an enum type with a type token, we know it's an enum
declared in an external Pulumi package. This should result in an enum
reference in our schema. Previously this would create a new enum
definition within our schema.

We also keep track of the Python types for these external enums, so that
we can deserialize the values back into the Python type.

Fixes https://github.com/pulumi/pulumi/issues/19297