Use RunPlugin for analyzer plugins (#19334)
This allows the engine to use language hosts to boot up policy packs. We
do this for all new language runtimes, i.e. not python and node, but
anything else with a "pulumi-language-X" plugin.

For python, node and any other policy runtime that doesn't have a
matching language runtime we fallback to the current behaviour of
looking for a "pulumi-analyzer-policy-X" plugin and running that like we
do currently. That is two args `[engine_address, "."]`, environment
variables like `PULUMI_STACK`/`PULUMI_CONFIG`, and with the working
directory set to the policy pack directory.

For new languages we run the policy by calling the languages RunPlugin
method with one argument (the engine address), no extra environment
variables, with kind set to "analyzer", and with the working directory
set to the main pulumi context directory (normally the top level pulumi
program directory).

New languages get the information that was historically passed via env
vars and working directory via `Handshake` instead.

The plan is to move nodejs and python to this model as well, but we'll
need to maintain some back-compatibility code in their runtimes instead.
Because that's non-trivial I want to defer that work till their policy
code is in pulumi/pulumi and can be well tested as we make that change.