codegen/python: don't use builtins.* prefix for docs (part 2)
In #19035 we started out writing the fully qualified types for python
in codegen, e.g. `builtins.str` instead of just `str`.

However this is not desired for docs, as it might confuse people. Don't
add the builtins prefix for docs generation.

This was partly fixed in https://github.com/pulumi/pulumi/pull/19359,
but I missed a spot. Fix that now.