Track YAML type inside the config system
This fixes the config values to track the YAML type they were marshalled from
so we don't do things like treat `"+111"` as an integer.

In fact with this we could probably get rid of the "adjustObjectValue" entirely and just use the type
from the YAML file, rather than heuristics to go from strings to typed values.

Fixes https://github.com/pulumi/pulumi/issues/19495.