Split analyzer handshake into handshake and stackconfigure (#19506)
This splits the analyzer handshake into two. We're doing this before we
hit the same issue we hit in providers where we used to call Configure
first to decide what the protocol to use with the provider was, and to
send configuration. But of course it didn't know what marshalling
options were safe for that configuration.

We _almost_ made the same mistake adding analyzer handshake where we
need some startup configuration, and so added it to handshake while
adding that method. But this is the same state we used to be in for
providers, so before anything actually uses this let's fix it.

Handshake is now _just_ for protocol setup. Currently analyzers can't
opt-in/out of anything but if we added integers/sets/etc we can now
safely opt-in to them in analyzers as they get updated.

StackConfigure (very open to better names here, but Configure was
already taken) then sends the ambient information about the stack being
analyzed to the analyzer (stack name, stack config, etc).

There's a bit of overlap here with GetAnalyzerInfo which returns if
Configure is supported, if we'd started with handshake that flag would
probably be on handshake. Plus there's the name oddity of Configure
already existing but that's for policy configuration which is separate
from the ambient stack info and config.