Elide `StackReference` outputs that fail to decrypt
`StackReference` is a resource offered by all core Pulumi SDKs, backed
internally by the built-in provider. `StackReference`s allow stacks to
read and refer to the outputs of other stacks. For example, given two
stacks A and B, here written in TypeScript:

```typescript
// a/index.ts

export = async () => {
  return {
    aString: "from-a",
  }
}

// b/index.ts

export = async () => {
  const aRef = new pulumi.StackReference("organization/a/dev");

  const aString = aRef.getOutput("aString")

  return {
    bString: pulumi.interpolate`from-b-and-${aString}`,
  }
}
```

Here, A produces a single output, `aString`, with the value `from-a`. B
creates a `StackReference` to A and loads the `aString` output, before
including it in its own output, yielding `bString: "from-b-and-from-a"`.

This all works well, and can in theory handle any output value that can
be serialized to state. Even secret values are supported, albeit with
one limitation -- the referencing stack (here B) must be able to
_decrypt_ the secrets of the referenced stack (here A). If both stacks
are e.g. using the same secrets manager (such as is the case with the
Pulumi service), everything is fine. If they are not, a decryption error
will likely be thrown, stopping the execution of the referencing stack.
This happens even if the referencing stack only uses non-secret values
(since presently, all outputs are loaded at once), causing issues such
as #11109.

In this case, we remove the requirement for `StackReference`s to be able
to decrypt secrets in referenced stacks. If a decryption fails, a dummy
value (that will likely cause errors if used) is returned, allowing
plain values to be shared freely. Additionally, in such cases, an `info`
diagnostic is emitted informing the user that secret values have been
elided, with an accompanying `debug` diagnostic (which is not visible by
default) holding the last encountered error message details.

Fixes #11109