[convert] Copy assets emitted from converter plugins to target directory of the generated project (#19500)
### Description

When `pulumi convert` talks to converter plugins, it tells them where
they should emit PCL files from which we bind a program and convert it.
If the converters happen to emit non-PCL files such as assets of the
program, we just ignore them. This PR addresses the issue by copying
non-PCL files from the source program directory (where PCL files were
emitted) into the target generated project directory so that the assets
stay part of the project.

This was implemented for conformance tests (specifically for the
`l2-resource-asset-archive` test) but not for `pulumi convert` so here
we implement it.

This might be used to support using local TF modules via
`pulumi-converter-terraform` where we could copy the contents of the
modules into the target project directory.

### Implementation

Extending `Copy` and `CopyDir` with a filter function `func(afero.File)
bool` to become `CopyWhen` and `CopyDirWhen`, then
- implementing `Copy` in terms of `CopyWhen` with a filter that always
passes
- implementing `CopyDir` in terms of `CopyDirWhen` with a filter that
always passes