implement plugin debugging in Go
Similar to NodeJS in https://github.com/pulumi/pulumi/pull/19517,
implement plugin debugging for Go plugins as well.  This adds a test
as well.

Unfortunately I couldn't figure out how to correctly make the plugin
continue correctly using the dap protocol, just how to shut it
down. Which makes the test a bit uglier, but it still shows that we
can communicate to the plugin with the debugger and shut the plugin
down correctly.

Co-authored-by: Eron Wright <eron@pulumi.com>