Fix hang in automation API operations (#19539)
Automation API starts a thread to watch the events log. If the
automation API call throws an error before the actual Pulumi operation
starts, we will never write a cancel event to the logs file, and the
thread hangs forever. Use a threading.Event to stop the polling loop to
ensure the thread always finishes.

---------

Co-authored-by: Thomas Gummerer <t.gummerer@gmail.com>