Add in venv backcompat for python when installing policy (#19570)
Fixes https://github.com/pulumi/pulumi/issues/19568

In https://github.com/pulumi/pulumi/pull/19477 we change policy
installation to use the language runtimes `InstallDependencies`
function. Unfortunately this doesn't auto-set the virtualenv option for
python, because for programs the logic has never been to auto-set it.
But policy packs _did_ autoset it to "venv" if it wasn't set and moving
to InstallDependencies lost that.

This adds a small special case to policy install to set the option, we
have very similar code in policy new.

---------

Co-authored-by: Julien <julien@caffeine.lu>