Add `DocLanguageHelper.GetResourceName`
This PR adds a new method to `DocLanguageHelper` (mirroring the existing
`GetFunctionName`), and an implementation for every language in this repo.

Motivated by https://github.com/pulumi/pulumi-service/issues/27266: The API we want to
support needs needs to provide an overview of whats in the package (resources and
functions). Concretely, that means it needs to return a list of modules, where each module
needs to itself have a list of the resources and functions it contains. Each element
returned (module, resource, function) needs to have a language appropriate name for each
language we support.

This PR is stacked on (and extremely similar to
https://github.com/pulumi/pulumi/pull/19569).