Enumerate supported CI systems deterministically (#19593)
When running a Pulumi operation, the Pulumi CLI is capable of detecting
various continuous integration and build systems and adding appropriate
metadata to the operation for analysis in e.g. Pulumi Cloud. For
instance, if one runs `pulumi up` in a Git repository, Pulumi will
attach metadata about the relevant commit hash, author, and so on. If
that operation is run as part of e.g. a GitHub Actions run, Pulumi will
capture metadata about that too.

This metadata can be really useful, but what if a user is using a CI
system or environment that Pulumi doesn't support or recognise? In this
case, users can manually set a series of `PULUMI_CI_*` environment
variables that Pulumi will consult in order to gather metadata.
Unfortunately, due to #7767, the enumeration of "CI detectors" is
actually non-deterministic. As a result, it's possible for a user to set
the `PULUMI_CI_*` environment variables in an environment that also
happens to set other variables the CLI looks for, and for Pulumi to pick
the wrong metadata (or at least, for it to pick metadata
inconsistently).

This change fixes this, making the enumeration of CI detectors
deterministic and placing the "generic" (`PULUMI_CI_*`-based) detector
at the front of the list. In theory we only support working when one set
of variables is defined anyway, and indeed, anyone with multiple sets
active will have been experiencing non-determinstic (read: broken)
behaviour to date, so this feels like a safe change.

Fixes #7767