implement plugin debugging in NodeJS (#19517)
Implement plugin debugging in NodeJS. Note that this required a small
change to the current debugging code for programs as well.
`--inspect-brk` always defaults to port 9229, even if another process is
already using that port for debugging. Therefore we need to print out
that port if it's not the default one, and also put it in the config.

Unfortunately no integration tests for the same reasons as in
https://github.com/pulumi/pulumi/pull/17156, and I still haven't found a
trick to do this in an integration test.

Based on top of https://github.com/pulumi/pulumi/pull/19511
/xref https://github.com/pulumi/home/issues/4061