Make pulumi install work for plugins (#19580)
This extends `pulumi install` which already works for programs and
policy packs to also work for plugins (i.e. folders with a
"PulumiPlugin.yaml" in them).

This allows us to use `pulumi install` in a number of places in the
integration tests to do npm/pip setup rather than dealing with those
commands directly.

This also means users can re-run installation for source based plugins
if needed.

---------

Co-authored-by: Thomas Gummerer <t.gummerer@gmail.com>