[convert] Ensure convert uses the output project directory as the working directory (#19582)
### Description

Converter plugins emit PCL that contains package descriptors with a
parameterization value. That parameterization sometimes include relative
paths that parameterized providers use to resolve assets e.g. resolving
paths to local TF modules. When resolving these relative paths, they
should be relative to output directory where the final Pulumi project is
generated, not the current working directory where `pulumi convert` is
running from which is the current behaviour today.

The current working directory where a package becomes parameterized from
is the working directory of the plugin context it was loaded from. This
PR changes the plugin context for the entirety of `pulumi convert` such
that it uses the output directory (where the pulumi generated project
will live) as the working directory of the process