Support refresh and import
Avoid contract violations for refreshes and import.

For refresh, we should only be refreshing the owning resource (skipping any views). The Read for the owning view should be responsible for updating any views.

For import, SameSteps are generated for the existing resources, including views. Loosen the contract to allow SameStep for view resources.