codegen: always initialize cache when initializing type (#19604)
This cache is used to cache intermediate results for type conversions
(see also https://github.com/pulumi/pulumi/pull/19422). However as these
functions can be called in parallel, if the cache is not initialized,
there is a possibility of a data race, checking for `t.cache == nil`
from one goroutine, while another writes to the variable.

We could fix this using a mutex, but this code is somewhat performance
critical, so it's better to just make sure to initialize this up front,
so we don't have to have a lock at every conversion.