remove downstream codegen tests
These tests have not been working for months, and we currently don't
have the engineering time allocated to fixing that (and fixing that is
a hard problem, because downstream depends on `pkg`, which we don't
consider a stable API interface, so they are going to keep breaking).

If we want to run these tests, we need to either not depend on `pkg`
downstream, or make `pkg` a stable API.  In the meantime, I think we
should not run them, as it just costs CI time, but more importantly,
makes every PR that touches codegen red, making the GitHub UI much
harder to work with.