stack select: better error in non-interactive mode
When `pulumi stack select` is run in non-interactive mode, and no
stack name is passed in as an argument to the command, we error out
with:

```
error: no stack selected; please use `pulumi stack select` or `pulumi stack init` to choose one
```

This is confusing to the user, as they are already running `pulumi
stack select`. Improve the error message here, to make this less
confusing to the user, and get them to either run the command in
interactive mode, or pass an argument to `pulumi stack select`.