Make the language tester a host itself (#19626)
As part of testing policy, but also if we want to do real provider
plugins later for provider testing, we need an engine host in the
conformance test process to pass to plugin start arguments.

There are two ways of doing this:

1. Add `EngineTarget` to `PrepareLanguageTestsRequest` and let the
language test process setup and run an engine host and pass it to the
conformance test process to use.
2. Make the conformance process itself an engine host. We can then just
write log messages to stderr. The language test process can then pick
them up by reading stderr and copying to `t.Log`. This means we don't
need the log de-duplication logic in each languages testcode.

I've gone with option two, as it means less is needed in each languages
test code (we currently duplicate the
`PULUMI_LANGUAGE_TEST_SHOW_FULL_OUTPUT` logic in every place).